requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
flake8==3.9.2
html5lib==1.1
idna==2.10
lxml==4.6.3
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
webencodings==0.5.1
